fix : 스프링 토큰 기반으로 회원 정보 찾게 변경, 그에 따라 Dto에 memberEmail항목 삭제 fix : Member.java에 getAge하는 방식 this.birthDate로 바꿈

+++ src/main/java/clubhouse/clubhouse/domain/application/dto/form/ApplyListResponseForm.java
@@ -7,12 +7,12 @@ import lombok.Data;
 public class ApplyListResponseForm {
 
     private String name;
-    private Long age;
+    private int age;
     private String school;
     private boolean isPass;
     private Long applicationId;
 
-    public ApplyListResponseForm(String name, Long age, String school, boolean is_pass, Long application_id) {
+    public ApplyListResponseForm(String name, int age, String school, boolean is_pass, Long application_id) {
         this.name = name;
         this.age = age;
         this.school = school;
